+++ HISTORY.rst
@@ -3,10 +3,17 @@ History
 =======
 
 
-0.1.5 (2023-XX-XX)
+0.1.6 (2023-XX-XX)
 -------------------------
 * Overhaul the ReadTheDocs documentation in the package
     * add examples and notes to each metric in package  
+
+0.1.6-alpha (2023-08-19)
+-------------------------
+* Fix Manney et al. 2011 implementation
+* Correct Manney 2011 method
+* Move old method to new metric: 'jet_core_identification_algorithm'
+* Update docs for Manney 2011 and sub-components
 
 
 0.1.5-beta (2023-08-16)